// Client/src/app/lib/interfaces/Employee.ts
export interface Employee {
    EmployeeId: number;
    FirstName: string;
    LastName: string;
    Email: string;
    DepartmentId: number;
    DepartmentName: string;
    HireDate: Date;
    CellPhone: string;
    Position: number;
    Salary: number;
    Balance: number;
    Address: string;
    Status: boolean;
    CreatedBy: number;
    CreatedDate: Date;
}